Change `show` to be isomorphic
This commit changes the `Show E164Number` implementation to always
return a string that when fed into `parseE164` will return the same
E164Number object.

The previous implementation would, for U.S. numbers, return a number
without +1. Unless this was added in, `parseE164` would have either
returned `Nothing` or an incorrect value. This code was relocated to a
new function called `human`.