Rename src to lib
Having thought about it, the src directory is too vague sounding; if
an application is broken into a library and an executable, it may be
in src and app (respectively), even though src used to house the
executable.

If a library exists with a small proof of concept, or demo executable,
that may also have the same src/app split.

This commit removes src for the convention lib/exe/test for library,
executable, and tests.